[#3] feat: 내비게이션 메뉴 컴포넌트 초안 작성, router 파일 분리
