+++ public/index.html
@@ -6,12 +6,6 @@
     <title>Document</title>
   </head>
   <body>
-    <!-- TODO: header 컴포넌트 분리 -->
-    <header id="header">
-      <button id="home">A</button>
-      <button id="about">B</button>
-    </header>
-
     <div id="root"></div>
   </body>
 </html>
