+++ src/index.ts
@@ -1,26 +1,12 @@
-import { RouterMap, createRouter } from './router'
+import { navMenu } from './NavMenu'
+import { router } from './router'
 
-const routerMap: RouterMap = {
-  home: {
-    path: '/home',
-    html: '<span>HOME</span>',
-  },
-  about: {
-    path: '/about',
-    html: '<span>ABOUT</span>',
-  },
+const $root = document.getElementById('root')
+
+if (!!$root) {
+  navMenu($root)
 }
-
-const router = createRouter(routerMap)
 
 window.addEventListener('DOMContentLoaded', (e) => {
   router.navigate(window.location.pathname.replace('/', ''))
-})
-
-// TODO: header 컴포넌트로 분리
-const $header = document.getElementById('header')
-
-$header?.addEventListener('click', (e) => {
-  const target = e.target as HTMLButtonElement
-  router.navigate(target.id)
 })
